Made Podcast card a separate component
